Screwing Everything Up
Was trying to make the map and image resizable, but couldn't manage
that. Then I was just trying to have the images all mapped like they
were before, but couldn't manage that. So really all I was able to do
was screw up quite a few of my previously working files. Cheers!